+++ tdbcli/filter.c
@@ -106,9 +106,9 @@ static tdb_item parse_term(const tdb *db,
     return item;
 }
 
-struct tdb_event_filter *parse_filter(const tdb *db,
-                                      const char *filter_expression,
-                                      int verbose)
+static struct tdb_event_filter *parse_filter(const tdb *db,
+                                             const char *filter_expression,
+                                             int verbose)
 {
     char *dup = strdup(filter_expression);
     char *expr = dup;
@@ -139,6 +139,108 @@ struct tdb_event_filter *parse_filter(const tdb *db,
     return filter;
 }
 
+static void apply_uuid(const char *uuidstr,
+                       tdb *db,
+                       struct tdb_event_filter *filter,
+                       uint32_t *num_invalid,
+                       uint32_t *num_missing)
+{
+    static uint8_t uuid[16];
+    uint64_t trail_id;
+    tdb_opt_value value = {.ptr = filter};
+
+    if (strlen(uuidstr) != 32 || tdb_uuid_raw(uuidstr, uuid)){
+        ++*num_invalid;
+        return;
+    }
+    if (tdb_get_trail_id(db, uuid, &trail_id)){
+        ++*num_missing;
+        return;
+    }
+    if (tdb_set_trail_opt(db, trail_id, TDB_OPT_EVENT_FILTER, value))
+        DIE("Could not set event filter");
+}
+
+static void apply_uuids_from_file(const char *fname,
+                                  tdb *db,
+                                  struct tdb_event_filter *filter,
+                                  uint32_t *num_uuids,
+                                  uint32_t *num_invalid,
+                                  uint32_t *num_missing)
+{
+    FILE *f;
+    char *line = NULL;
+    size_t n = 0;
+
+    if (!(f = fopen(fname, "r")))
+        DIE("Could not read UUIDs from %s\n", fname);
+
+    while (getline(&line, &n, f) != -1){
+        ++*num_uuids;
+        line[strlen(line) - 1] = 0;
+        apply_uuid(line, db, filter, num_invalid, num_missing);
+    }
+
+    fclose(f);
+    free(line);
+}
+
+static void apply_uuids(tdb *db,
+                        struct tdb_event_filter *filter,
+                        struct tdbcli_options *opt)
+{
+    char *dup = strdup(opt->uuids);
+    uint32_t num_uuids = 0;
+    uint32_t num_invalid = 0;
+    uint32_t num_missing = 0;
+
+    if (dup[0] == '@')
+        apply_uuids_from_file(&dup[1],
+                              db,
+                              filter,
+                              &num_uuids,
+                              &num_invalid,
+                              &num_missing);
+    else{
+        char *uuidstr;
+        while ((uuidstr = strsep(&dup, ","))){
+            ++num_uuids;
+            apply_uuid(uuidstr, db, filter, &num_invalid, &num_missing);
+        }
+    }
+    if (opt->verbose)
+        fprintf(stderr,
+                "Found %u UUIDs: %u selected, %u missing, %u invalid.\n",
+                num_uuids,
+                num_uuids - (num_invalid + num_missing),
+                num_missing,
+                num_invalid);
+    free(dup);
+}
+
+struct tdb_event_filter *apply_filter(tdb *db, struct tdbcli_options *opt)
+{
+    struct tdb_event_filter *f = NULL;
+    tdb_opt_value value;
+
+    if (opt->filter_arg)
+        f = parse_filter(db, opt->filter_arg, opt->verbose);
+
+    if (opt->uuids){
+        /* all and none are never freed, not a biggie */
+        if (f)
+            apply_uuids(db, f, opt);
+        else
+            apply_uuids(db, tdb_event_filter_new_match_all(), opt);
+        value.ptr = tdb_event_filter_new_match_none();
+    }else
+        value.ptr = f;
+
+    if (tdb_set_opt(db, TDB_OPT_EVENT_FILTER, value))
+        DIE("Could not set event filter");
+
+    return f;
+}
 
 #ifdef TEST_FILTER
 int main(int argc, char **argv)